+++ tests/testthat/test_rwtbl_helpers.R
@@ -58,9 +58,8 @@ test_that("`rwtbl_var_to_slot()` errors correctly", {
   )
 })
 
-rwa <- rwd_agg(read.csv(
-  system.file("extdata/rwd_agg_files/passing_aggs.csv", package = "RWDataPlyr"),
-  stringsAsFactors = FALSE)
+rwa <- read_rwd_agg(
+  system.file("extdata/rwd_agg_files/passing_aggs.csv", package = "RWDataPlyr")
 )
 
 test_that("`rwtbl_var_to_slot()` works", {